Full CRUD and forms on Product

+++ ex-2/app/models/product/Product.scala
@@ -2,7 +2,7 @@ package models.product
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Product(id: Long, name: String, description: String, category: Long, producer: Long, price: Double)
+case class Product(id: Long, name: String, description: String, category: Long, producer: Long, price: Int)
 
 object Product {
   implicit val productFormat: OFormat[Product] = Json.format[Product]
